requirements.txt
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
debtagshw==0.1
defer==1.0.6
dirspec==13.10
django-cors-headers==1.0.0
django-easyregistration==0.1
django-forms-bootstrap==3.0.1
django-oauth2-provider==0.2.6.1
djangorestframework==3.0.2
html5lib==0.999
httpie==0.8.0
httplib2==0.8
lxml==3.3.3
numpy==1.8.2
oauthlib==0.6.1
oneconf==0.3.7
pdfminer==20140328
pexpect==3.1
piston-mini-client==0.7.5
psutil==1.2.1
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5
python-augeas==0.4.1
python-debian==0.1.21-nmu2ubuntu2
pyxdg==0.25
reportlab==3.0
requests==2.2.1
sessioninstaller==0.0.0
shortuuid==0.4.2
six==1.5.2
software-center-aptd-plugins==0.0.0
urllib3==1.7.1
vboxapi==1.0
wsgiref==0.1.2
zope.interface==4.0.5
